+++ react/typings/render.d.ts
@@ -1,0 +1,3 @@
+module 'render' {
+  export const withRuntimeContext
+}