+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- OrderNow metric on splunk.
+
 ## [0.8.0] - 2019-04-12
 
 ### Added